try to avoid casting to int
git-svn-id: https://svn.apache.org/repos/asf/poi/trunk@1866933 13f79535-47bb-0310-9956-ffa450edef68

+++ src/examples/src/org/apache/poi/hssf/view/SVSheetTable.java
@@ -182,7 +182,7 @@ public class SVSheetTable extends JTable {
       Row row = sheet.getRow(i - sheet.getFirstRowNum());
       if (row != null) {
         short h = row.getHeight();
-        int height = (int)Math.round(Math.max(1., h / (res / 70. * 20.) + 3.));
+        int height = Math.toIntExact(Math.round(Math.max(1., h / (res / 70. * 20.) + 3.)));
         System.out.printf("%d: %d (%d @ %d)%n", i, height, h, res);
         setRowHeight(i, height);
       }
